# optimize-put-float-ascii-method